# backend/src/openapi.yaml
swagger: "2.0"
info:
  description: "This is an API for a stock-market simulation game. Let users bet with fake money on real market values. Find the code in our [GitHub Repo](https://github.com/the-philip-baumann/mock-stock)"
  version: "1.0"
  title: "Mock-Stock"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
securityDefinitions:
  cookieAuth:
    type: apiKey
    in: header
    name: Set-Cookie
    description: Session Cookies etc.
paths:
  /stocks:
    get:
      tags:
      - "stocks"
      summary: "Gets list of stocks"
      description: "Gets the stock values from the real market."
      operationId: "getAllStock"
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "top"
        description: "Gets only the top so much stocks"
        required: true
        type: "integer"
      - in: "query"
        name: "skip"
        description: "Skips an amount of stocks and then returns the amount of stocks defined in top"
        required: false
        type: "integer"
      responses:
        "200":
          description: "Stocks found"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SimpleStock"
        "404":
          description: "No stonks today"
  /stocks/{stock}:
    get:
      tags:
      - "stocks"
      summary: "Get the values of provided stock"
      description: "Gets the stock values from the real market."
      operationId: "getStock"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "stock"
        description: "The Stock name e.g. AAPL from which we want the data."
        required: true
        type: "string"
      responses:
        "200":
          description: "Stock found"
          schema:
            $ref: "#/definitions/Stock"
        "404":
          description: "Stock not found"
    put:
      tags:
      - "stocks"
      summary: "Buy or sell stocks"
      description: "Buy or sell an amount of Stocks"
      operationId: "buyStock"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "stock"
        description: "Stock to buy"
        required: true
        type: "string"
      - in: "body"
        name: "amount"
        description: "Amount to sell if negativ, amount to buy if positiv"
        required: true
        schema:
          type: "integer"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "Bought or sold stock"
          schema:
            $ref: "#/definitions/Stock"
        "403":
          description: "Not logged in"
        "404":
          description: "Stock not found"
        "405":
          description: "Too poor to buy those stocks *sad face*"
...
definitions:
  Stock:
    type: "object"
    properties:
      id:
        type: "string"
        example: "AAPL"
        description: "ID of stock"
      value:
        type: "number"
        example: 155.3456
        description: "Value of stock in USD"
      name:
        type: "string"
        example: "Apple Inc."
        description: "The Name of the Company"
      amount:
        type: "integer"
        format: "int64"
        example: 3
        description: "The amount of this stock someone holds."
      valueHistory:
        type: "array"
        items:
          type: "number"
        example: [170.2, 120.22, 140.00]
        description: "The History of the score. Duration over 5 days with 1 hour intervals, first value is newest."
  SimpleStock:
    type: "object"
    properties:
      id:
        type: "string"
        example: "AAPL"
        description: "ID of stock"
      value:
        type: "number"
        example: 155.3456
        description: "Value of stock in USD"
      name:
        type: "string"
        example: "Apple Inc."
        description: "The Name of the Company"
      amount:
        type: "integer"
        format: "int64"
        example: 3
        description: "The amount of this stock someone holds."
      change:
        type: "number"
        example: 2.93
        description: "How much the stock changed over the last hour"
    description: "Stock Dto with less overhead for List views."
...